+++ docs/metro/js/metro.js
@@ -4150,6 +4150,7 @@ var Accordion = {
         return this;
     },
     options: {
+        mode: "default",
         duration: METRO_ANIMATION_DURATION,
         oneFrame: true,
         showActive: true,
@@ -4184,6 +4185,10 @@ var Accordion = {
         var frame_to_open;
 
         element.addClass("accordion");
+
+        if (o.mode.toLowerCase() !== "default") {
+            element.addClass(o.mode);
+        }
 
         if (active.length === 0) {
             frame_to_open = frames[0];
@@ -5052,6 +5057,114 @@ var Audio = {
 };
 
 Metro.plugin('audio', Audio);
+
+// Source: js/plugins/bottom-sheet.js
+var BottomSheet = {
+    init: function( options, elem ) {
+        this.options = $.extend( {}, this.options, options );
+        this.elem  = elem;
+        this.element = $(elem);
+        this.toggle = null;
+
+        this._setOptionsFromDOM();
+        this._create();
+
+        return this;
+    },
+
+    options: {
+        mode: "list",
+        toggle: null,
+        onOpen: Metro.noop,
+        onClose: Metro.noop,
+        onBottomSheetCreate: Metro.noop
+    },
+
+    _setOptionsFromDOM: function(){
+        var that = this, element = this.element, o = this.options;
+
+        $.each(element.data(), function(key, value){
+            if (key in o) {
+                try {
+                    o[key] = JSON.parse(value);
+                } catch (e) {
+                    o[key] = value;
+                }
+            }
+        });
+    },
+
+    _create: function(){
+        var that = this, element = this.element, o = this.options;
+
+        this._createStructure();
+        this._createEvents();
+
+        Utils.exec(o.onBottomSheetCreate, [element], element[0]);
+    },
+
+    _createStructure: function(){
+        var that = this, element = this.element, o = this.options;
+
+        element.addClass(o.mode+"-list");
+
+        if (Utils.isValue(o.toggle) && $(o.toggle).length > 0) {
+            this.toggle = $(o.toggle);
+        }
+    },
+
+    _createEvents: function(){
+        var that = this, element = this.element, o = this.options;
+
+        if (Utils.isValue(this.toggle)) {
+            this.toggle.on(Metro.events.click, function(){
+                that.toggleView();
+            });
+        }
+
+        element.on(Metro.events.click, "li", function(){
+            that.close();
+        });
+    },
+
+    isOpen: function(){
+        return this.element.hasClass("opened");
+    },
+
+    open: function(mode){
+        var element = this.element, o = this.options;
+
+        if (Utils.isValue(mode)) {
+            element.removeClass("list-style grid-style").addClass(mode+"-style");
+        }
+
+        this.element.addClass("opened");
+        Utils.exec(o.onOpen, [element], element[0]);
+    },
+
+    close: function(){
+        var element = this.element, o = this.options;
+
+        element.removeClass("opened");
+        Utils.exec(o.onClose, [element], element[0]);
+    },
+
+    toggleView: function(){
+        if (this.isOpen()) {
+            this.close();
+        } else {
+            this.open();
+        }
+    },
+
+    changeAttribute: function(attributeName){
+
+    },
+
+    destroy: function(){}
+};
+
+Metro.plugin('bottomsheet', BottomSheet);
 
 // Source: js/plugins/button-group.js
 var ButtonGroup = {