+++ CHANGELOG.md
@@ -7,7 +7,11 @@
 + [x] Notify: add methods `onNotifyCreate`, `onAppend`
 + [x] Input: fix custom search button click
 + [x] Calendar picker: add attribute `data-null-value`. If this attribute false and value empty, used current date
-+ [x] Calendar picker: fix for null value, issue #1217  
++ [x] Calendar picker: fix for null value, issue #1217
++ [x] Accordion: add attribute `data-mode='material'`
++ [x] Mif: add new icons
++ [x] Bottom navigation: add new component  
++ [x] Bottom sheet: add new component  
 
 ### 4.2.23
 + [x] Select: fix native `onchange` event triggering, issue #1198